+++ lib/config.dart
@@ -25,6 +25,22 @@ class FlutterContactsConfig {
   /// notes.
   bool includeNotesOnIos13AndAbove = false;
 
+  /// Non-visible contacts on Android are contacts that are not part of a group,
+  /// and are excluded by default.
+  ///
+  /// See https://stackoverflow.com/questions/28665587/what-does-contactscontract-contacts-in-visible-group-mean-in-android
+  bool includeNonVisibleOnAndroid = false;
+
+  /// Return unified contacts instead of raw contacts.
+  ///
+  /// On both iOS and Android there is a concept of raw and unified contacts. A
+  /// single person might have two raw contacts (for example from Gmail and from
+  /// iCloud) but will be merged into a single view called a unified contact. In
+  /// a contact app you typically want unified contacts, so this is what's
+  /// returned by default. When this is false, raw contacts are returned
+  /// instead.
+  bool returnUnifiedContacts = true;
+
   /// The vCard version to use when exporting to VCard. V4 is the most current,
   /// but V3 is the most commonly supported.
   VCardVersion vCardVersion = VCardVersion.v3;